+++ backend/app/models/plant.rb
@@ -1,5 +1,6 @@
 class Plant < ApplicationRecord
   include RestorableInterface
+  include Auditable
   belongs_to :deleted_by, class_name: "User", optional: true
   belongs_to :lote
   belongs_to :club, optional: true